Added DI for repository via interfaces
